requirements.txt
django
djangorestframework
django-sonar
django-allauth
drf-yasg
django-crispy-forms